Until now a `Range` could only step by an integer value, even if its elements had other types, e.g. doubles.
This made stepping over ranges like `0.1..0.9`strange, as that resulted in a single element.

Also, stepping over `ObjectRange` instances (like the above one) called the `next`/`previous` method reflectively, even if the range had a numeric type (the most common case).
This made it ~10x slower than e.g. `IntRange`.

The new implementation changed the `step` type to `Number` and unified the way stepping is done in
`ObjectRange` and `IntRange`, making them as fast as the original `IntRange` for numeric cases.

Also, `IntRange` treats the overflow issue by storing the intermediary value in a `Long` directly.

At the boundaries, the `next` and `previous` methods may now wrap to the other extreme - or return null.

-------------------

I tested the speed via the following trivial method:

```Groovy
void testSpeed() {
    for (def i = 0; i < 10; i++)
        run()
}

private void run() {
    def time = System.currentTimeMillis()
    def size = 0G
    for (def i = 0; i < 100; i++) { // changed to `i = 0G` when testing ObjectRange
        for (def j = 0; j < 100; j++) {
            def range = i..j
            def inverseRange = j..i
            for (def k = 1; k < 100; k++) {
                size += range.step(k).size()
                size += range.step(-k).size()

                size += inverseRange.step(k).size()
                size += inverseRange.step(-k).size()
            }
        }
    }
    println "size = ${size} in ${(System.currentTimeMillis() - time) / 1000.0}s"
}
```

resulting in
```
* Old impl for ObjectRange: `66.30s`
* New impl for ObjectRange: `5.52s` (12x faster)
```
and
```
* Old impl for IntRange: `4.54s`
* New impl for IntRange: `4.52s` (same speed as before)
```