@paplorinc 2015-09-17
Should play well with https://github.com/apache/incubator-groovy/pull/109