@blackdrag 2015-09-18
your change currently has two obvious problems. (1) returning Collections.emptyList() is only ok, if you say, that the result cannot be modified. The old implementation did not have this. You even had to change a test for this. (2) You changed step(int) to step(Number). This is not binary compatible. Java programs using Range directly will have to recompile. They will do so just fine, since Java will do the boxing, but an old library will not work with these new Ranges