@paplorinc 2015-09-18
Hello blackDrag,

1) I will document it, if you think it's necessary, though I think it was already specified that it returned an empty list - and not that it returned a new list. Also, why would people modify a list, returned by an empty range?

2) I will add a deprecated delegate method, if you think there will be people who want the latest Groovy, with and older lib, using `step`.