@blackdrag 2015-09-18
An empty list can still be a list that gets elements added.  One more thing... can you show me the benchmark code you did for ObjectRange as well?