@paplorinc 2015-09-18
Yes, I changed the description to `immutable` empty list.

The `ObjectRange` is described in the above description in the comment: `// changed to `i = 0G` when testing ObjectRange`