@PascalSchumacher 2015-09-26
@paplorinc Your changes crash the build http://ci.groovy-lang.org/viewLog.html?buildId=27049&tab=buildResultsDiv&buildTypeId=Groovy_Jdk7Build