@paplorinc 2015-09-26
Thanks, will rerun it locally (can't view the TC build).
Could we enable CI trigger for pull requests also, as it's done for other OS projects (e.g. https://github.com/google/guava/pull/2164#partial-pull-merging)?