@PascalSchumacher 2015-09-26
@paplorinc You can view the tc build if you login as "guest" with a blank password.

CI trigger for pull request would be very helpful (we had this before the move to apache), but I was told it's not possible with teamcity and the read-only apache github mirror.