@paplorinc 2015-09-26
Thanks, the problem was, that in case of enums, at the boundaries, the `previous` and `next` methods have wrapped to the other extreme.
I.e. `next` for the last element returned the first one (I didn't know this was legal behavior).
Added a separate test case for this and for `null` values also.