@paplorinc 2015-09-27
http://ci.groovy-lang.org/viewLog.html?buildId=27169&tab=buildResultsDiv&buildTypeId=Groovy_Jdk7Build