@tkruse 2016-06-28
Since #143 was merged, this PR should be rebased, and some conflicts can be expected.