@paplorinc 2016-06-28
Thanks, but I will rather abandon it at this stage.