@tkruse 2016-06-28
Possibly at least the increment/decrement optimization for numbers could be kept