@paulk-asert 2016-06-29
I copied out a few of the additional tests you added to cover existing functionality. Also, PRs #142 and #143 have been merged recently. Any chance you can rework this PR to be just the latest changes required for the Number step functionality?