@paulk-asert 2016-06-29
Just for fun, I had a look at what a utility method outside of Range might look like:
https://github.com/apache/groovy/pull/359
I am wondering whether that might limit added complexity to the Range class? I'd be interested in what others think.