@paplorinc 2016-06-29
Revived in https://github.com/apache/groovy/pull/360, please review.
Is it ok that the new `Gradle 3` crashes the build on the `java 6` Travis node?